requirements-airflow.txt
pysus==0.5.0
geopandas==0.14.0
pandas==2.2.0
pyarrow==15.0.0
psycopg2-binary==2.9.9
requests==2.31.0